+++ src/ReadableAtom_/index.d.ts
@@ -1,5 +1,6 @@
 import type { ReadableAtom } from 'nanostores'
 export interface ReadableAtom_<Value> extends ReadableAtom<Value> {
+	():Value
 	readonly _:Value
 	readonly $:Value
 }