+++ src/WritableAtom_/index.d.ts
@@ -1,6 +1,8 @@
 import type { WritableAtom } from 'nanostores'
 import { ReadableAtom_ } from '../ReadableAtom_'
 export interface WritableAtom_<Value> extends ReadableAtom_<Value>,WritableAtom<Value> {
+	():Value
+	(_:Value):void
 	_:Value
 	$:Value
 }