+++ src/_private/index.d.ts
@@ -1,0 +1,5 @@
+import { ReadableAtom } from 'nanostores'
+import { ReadableAtom_ } from '../ReadableAtom_'
+import { WritableAtom_ } from '../WritableAtom_'
+export declare function readable_fn_<Value>(atom:ReadableAtom):ReadableAtom_<Value>
+export declare function writable_fn_<Value>(atom:ReadableAtom):WritableAtom_<Value>