+++ src/computed_/index.js
@@ -1,15 +1,7 @@
 import { computed as _computed_ } from 'nanostores'
+import { readable_fn_ } from '../_private/index.js'
 export const computed_ = (stores, cb)=>{
-	const computed = _computed_(stores, cb)
-	Object.defineProperties(computed, Object.getOwnPropertyDescriptors({
-		get _() {
-			return computed.get()
-		},
-		get $() {
-			return computed.get()
-		},
-	}))
-	return computed
+	return readable_fn_(_computed_(stores, cb))
 }
 export {
 	computed_ as computed$