+++ src/neq_atom_/index.js
@@ -1,14 +1,14 @@
-import { onSet } from 'nanostores'
-import { atom_ } from '../atom_/index.js'
+import { atom as _atom_, onSet } from 'nanostores'
+import { writable_fn_ } from '../_private/index.js'
 /** @type {import('./index.d.ts').neq_atom_} */
 export const neq_atom_ = (initialValue)=>{
-	const atom = atom_(initialValue)
-	decorate_neq_atom_(atom)
-	return atom
+	const _atom = _atom_(initialValue)
+	decorate_neq_atom_(_atom)
+	return writable_fn_(_atom)
 }
 export function decorate_neq_atom_(atom) {
 	return onSet(atom, ({ newValue, abort })=>{
-		if (atom._ == newValue) abort()
+		if (atom.get() == newValue) abort()
 	})
 }
 export {