+++ src/setter_computed_/index.test.ts
@@ -11,26 +11,29 @@ test.after(()=>{
 	clock.uninstall()
 })
 test('setter_computed_()._', ()=>{
-	let letter$ = atom_('a')
-	let number$ = atom_(0)
-	const a = setter_computed_([letter$, number$], ([letter, number], set)=>{
+	let letter_ = atom_('a')
+	let number_ = atom_(0)
+	const a_ = setter_computed_([letter_, number_], ([letter, number], set)=>{
 		run(async ()=>{
 			set(`${letter}-${String(number * 2)}`)
 		})
 	})
-	is(a._, 'a-0')
-	is(a.$, 'a-0')
-	is(a.get(), 'a-0')
-	letter$._ = 'b'
+	is(a_(), 'a-0')
+	is(a_._, 'a-0')
+	is(a_.$, 'a-0')
+	is(a_.get(), 'a-0')
+	letter_('b')
 	clock.runAll()
-	is(a._, 'b-0')
-	is(a.$, 'b-0')
-	is(a.get(), 'b-0')
-	number$._ = 2
+	is(a_(), 'b-0')
+	is(a_._, 'b-0')
+	is(a_.$, 'b-0')
+	is(a_.get(), 'b-0')
+	number_(2)
 	clock.runAll()
-	is(a._, 'b-4')
-	is(a.$, 'b-4')
-	is(a.get(), 'b-4')
+	is(a_(), 'b-4')
+	is(a_._, 'b-4')
+	is(a_.$, 'b-4')
+	is(a_.get(), 'b-4')
 })
 test('setter_computed_|single & multiple (array) dependency stores', ()=>{
 	const i0_ = atom_(0)
@@ -38,11 +41,17 @@ test('setter_computed_|single & multiple (array) dependency stores', ()=>{
 	const single_ = setter_computed_(i0_, (args, set)=>{
 		set(args)
 	})
+	equal(single_(), 0)
 	equal(single_._, 0)
+	equal(single_.$, 0)
+	equal(single_.get(), 0)
 	const multi_ = setter_computed_([i0_, i1_], (args, set)=>{
 		set(args)
 	})
+	equal(multi_(), [0, 1])
 	equal(multi_._, [0, 1])
+	equal(multi_.$, [0, 1])
+	equal(multi_.get(), [0, 1])
 })
 test('setter_computed_|child atoms|cb does not set', ()=>{
 	const atom = atom_(0)
@@ -50,11 +59,14 @@ test('setter_computed_|child atoms|cb does not set', ()=>{
 		return
 	})
 	const child_ = computed_(setter_computed, _=>_)
-	const child__listen_a:number[] = []
-	child_.listen(_=>child__listen_a.push(_))
-	atom._ = 1
+	const child_listen_a:number[] = []
+	child_.listen(_=>child_listen_a.push(_))
+	atom(1)
+	equal(child_(), undefined)
 	equal(child_._, undefined)
-	equal(child__listen_a, [])
+	equal(child_.$, undefined)
+	equal(child_.get(), undefined)
+	equal(child_listen_a, [])
 })
 test('setter_computed_|cb returns sets same value', ()=>{
 	const atom = atom_<Object>()
@@ -62,15 +74,21 @@ test('setter_computed_|cb returns sets same value', ()=>{
 		set(atom)
 	})
 	const child_ = computed_(setter_computed, _=>_)
-	const child__listen_a:Object[] = []
-	child_.listen(_=>child__listen_a.push(_))
+	const child_listen_a:Object[] = []
+	child_.listen(_=>child_listen_a.push(_))
 	const val = {}
-	atom._ = val
+	atom(val)
+	equal(child_(), val)
 	equal(child_._, val)
-	equal(child__listen_a, [val])
-	atom._ = val
+	equal(child_.$, val)
+	equal(child_.get(), val)
+	equal(child_listen_a, [val])
+	atom(val)
+	equal(child_(), val)
 	equal(child_._, val)
-	equal(child__listen_a, [val, val])
+	equal(child_.$, val)
+	equal(child_.get(), val)
+	equal(child_listen_a, [val, val])
 })
 test('setter_computed_|cb returns set then no set', ()=>{
 	const atom = atom_<Object>()
@@ -82,14 +100,20 @@ test('setter_computed_|cb returns set then no set', ()=>{
 		}
 	})
 	const child_ = computed_(setter_computed, _=>_)
-	const child__listen_a:Object[] = []
-	child_.listen(_=>child__listen_a.push(_))
+	const child_listen_a:Object[] = []
+	child_.listen(_=>child_listen_a.push(_))
 	const val = {}
-	atom._ = val
+	atom(val)
+	equal(child_(), val)
 	equal(child_._, val)
-	equal(child__listen_a, [val])
-	atom._ = val
+	equal(child_.$, val)
+	equal(child_.get(), val)
+	equal(child_listen_a, [val])
+	atom(val)
+	equal(child_(), val)
 	equal(child_._, val)
-	equal(child__listen_a, [val])
+	equal(child_.$, val)
+	equal(child_.get(), val)
+	equal(child_listen_a, [val])
 })
 test.run()