+++ src/split_atom_/index.js
@@ -1,17 +1,8 @@
 import { atom } from 'nanostores'
+import { readable_fn_ } from '../_private/index.js'
 export const split_atom_ = initialValue=>{
 	const split_atom = atom(initialValue)
-	const { set } = split_atom
-	delete split_atom.set
-	Object.defineProperties(split_atom, Object.getOwnPropertyDescriptors({
-		get _() {
-			return split_atom.get()
-		},
-		get $() {
-			return split_atom.get()
-		},
-	}))
-	return [split_atom, set]
+	return [readable_fn_(split_atom), split_atom.set]
 }
 export {
 	split_atom_ as split_atom$