+++ src/split_atom_/index.test.ts
@@ -1,14 +1,16 @@
 import { test } from 'uvu'
 import { is } from 'uvu/assert'
 import { split_atom_ } from '../index.js'
-test('split_atom_._', ()=>{
-	const [a$, set] = split_atom_(1)
-	is(a$._, 1)
-	is(a$.$, 1)
-	is(a$.get(), 1)
+test('split_atom_()()', ()=>{
+	const [a_, set] = split_atom_(1)
+	is(a_(), 1)
+	is(a_._, 1)
+	is(a_.$, 1)
+	is(a_.get(), 1)
 	set(2)
-	is(a$._, 2)
-	is(a$.$, 2)
-	is(a$.get(), 2)
+	is(a_(), 2)
+	is(a_._, 2)
+	is(a_.$, 2)
+	is(a_.get(), 2)
 })
 test.run()